Fixed that Get-PnPFile -AsListItem returns all properties of the file and not just the default 3. Added a generic method for converting a listitem to a PSObject.
